requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.2
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Werkzeug==3.0.1
SQLAlchemy==1.4.51
email-validator==2.1.0.post1
python-dotenv==1.0.0
bcrypt==4.1.1
